topnav update

+++ query.php
@@ -37,11 +37,22 @@
 		header('location:borrow.php');
 	}
 
+	//if request again is clicked on borrowed
+	else if(isset($_GET["request_again"])) {
+		$bid = $_GET['request_again'];
+		$to_user = $_GET['to_user'];
+		$sql = "delete from requests where bid='" . $bid . "' and from_user='" . $_SESSION['user_id'] . "' and to_user='" . $to_user . "'";
+		mysqli_query($link,$sql);
+		$sql = "insert into requests values('" . $bid . "','" . $_SESSION['user_id'] . "','" . $to_user . "',0,1,0)";
+		mysqli_query($link,$sql);
+		header('location:borrow.php');
+	}
+
 	//if return is clicked on borrow
 	else if(isset($_GET["return"])) {
 		$bid = $_GET['return'];
 		$to_user = $_GET['to_user'];
-		$sql = "delete from requests where bid='" . $bid . "' and to_user='" . $to_user . "'";
+		$sql = "delete from requests where bid='" . $bid . "' and from_user='" . $_SESSION['user_id'] . "' and to_user='" . $to_user . "'";
 		mysqli_query($link,$sql);
 		header('location:borrow.php');
 	}
@@ -50,7 +61,7 @@
     else if(isset($_GET["accept"])) {
         $bid=$_GET['accept'];
 		$from_user = $_GET['from_user'];
-        $sql="UPDATE requests SET sn=1,rn=0,status=1 WHERE bid='$bid' and from_user='$from_user'";
+        $sql="UPDATE requests SET sn=1,rn=0,status=1 WHERE bid='$bid' and from_user='$from_user' and to_user='" . $_SESSION['user_id'] . "'";
         mysqli_query($link,$sql);
         header('location:lent.php');
     }
@@ -59,7 +70,7 @@
     else if(isset($_GET["decline"])) {
         $bid=$_GET['decline'];
 		$from_user = $_GET['from_user'];
-        $sql="UPDATE requests SET sn=1,rn=0,status=2 WHERE bid='$bid' and from_user='$from_user'";
+        $sql="UPDATE requests SET sn=1,rn=0,status=2 WHERE bid='$bid' and from_user='$from_user' and to_user='" . $_SESSION['user_id'] . "'";
         mysqli_query($link,$sql);
         header('location:lent.php');
     }
@@ -68,7 +79,7 @@
     else if(isset($_GET["cancel"])) {
         $bid=$_GET['cancel'];
 		$from_user = $_GET['from_user'];
-        $sql="UPDATE requests r SET r.sn=0,r.rn=1,status=0 WHERE bid='$bid' and from_user='$from_user'";
+        $sql="UPDATE requests r SET r.sn=0,r.rn=1,status=0 WHERE bid='$bid' and from_user='$from_user' and to_user='" . $_SESSION['user_id'] . "'";
         mysqli_query($link,$sql);
         header('location:lent.php');
     }
